requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
python-dotenv
chromadb
tiktoken
fastapi
uvicorn
pydantic
requests